WIP refactor working
