+++ baseapp/testapp.go
@@ -1,24 +1,23 @@
-package testapp
+package baseapp
 
 import (
 	abci "github.com/tendermint/abci/types"
 
-	bam "github.com/cosmos/cosmos-sdk/baseapp"
-	x "github.com/cosmos/cosmos-sdk/baseapp/testapp/x"
+	"github.com/cosmos/cosmos-sdk/baseapp/testtx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // TestApp wraps BaseApp with helper methods,
 // and exposes more functionality than otherwise needed.
 type TestApp struct {
-	*bam.BaseApp
+	*BaseApp
 
 	// These get set as we receive them.
 	*abci.ResponseBeginBlock
 	*abci.ResponseEndBlock
 }
 
-func NewTestApp(bapp *bam.BaseApp) *TestApp {
+func NewTestApp(bapp *BaseApp) *TestApp {
 	app := &TestApp{
 		BaseApp: bapp,
 	}
@@ -78,14 +77,14 @@ func (tapp *TestApp) RunDeliverTx(tx sdk.Tx) sdk.Result {
 // run tx through CheckTx of TestApp
 // NOTE: Skips authentication by wrapping msg in TestTx{}.
 func (tapp *TestApp) RunCheckMsg(msg sdk.Msg) sdk.Result {
-	var tx = x.TestTx{msg}
+	var tx = testtx.TestTx{msg}
 	return tapp.RunCheckTx(tx)
 }
 
 // run tx through DeliverTx of TestApp
 // NOTE: Skips authentication by wrapping msg in TestTx{}.
 func (tapp *TestApp) RunDeliverMsg(msg sdk.Msg) sdk.Result {
-	var tx = x.TestTx{msg}
+	var tx = testtx.TestTx{msg}
 	return tapp.RunDeliverTx(tx)
 }
